Move Folder to bitwarden::vault
