+++ crates/bitwarden/src/vault/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "mobile")]
+mod folder;
+
+#[cfg(feature = "mobile")]
+pub use folder::*;
